Added clear line, Process all Exits
clear_to_end_of_current_line() clears 50 characters of what's printed on screen

We now loop through all exits for the room and act upon the correct one instead of hardcoded.

Prompt now clears after successful room change.

+++ src/ui/tui.zig
@@ -43,7 +43,7 @@ pub const Tui = struct {
         _ = try stdout.write(RESTORE_CURSOR);
     }
 
-    pub fn move_cursor_direction(_: *const Tui, comptime direction: Direction, count: u16) !void {
+    pub fn move_cursor_direction(_: *const Tui, comptime direction: Direction, count: usize) !void {
         const fmt = comptime switch (direction) {
             .up => MOVE_CURSOR_UP,
             .down => MOVE_CURSOR_DOWN,
